+++ lua/obsidian/lsp/handlers/_references.lua
@@ -2,6 +2,7 @@ local util = require "obsidian.util"
 local log = require "obsidian.log"
 local api = require "obsidian.api"
 local search = require "obsidian.search"
+local parse_block_id = require "obsidian.parse.block_id"
 
 ---@param match obsidian.BacklinkMatch
 ---@return lsp.Location
@@ -30,10 +31,7 @@ local function tag_loc_to_lsp_location(tag_loc)
   }
 end
 
----@type table<obsidian.search.RefTypes, fun(_: string, callback: fun(locs: lsp.Location[]))|?>
-local handlers = {}
-
-handlers.Markdown = function(link, callback)
+local function handle_note_ref(link, callback)
   local location = util.parse_link(link)
   assert(location, "failed to parse link")
 
@@ -47,17 +45,41 @@ handlers.Markdown = function(link, callback)
   local anchor_link
   location, anchor_link = util.strip_anchor_links(location)
 
-  local opts = { anchor = anchor_link, block = block_link, refs = { location } }
+  local opts = { anchor = anchor_link, block = block_link }
 
-  search.find_backlinks_async(nil, function(backlink_matches)
-    callback(vim.iter(backlink_matches):map(backlink_to_lsp_location):totable())
-  end, opts)
+  local function find_backlinks(note)
+    if note == nil then
+      opts.refs = { location }
+    end
+    search.find_backlinks_async(note, function(backlink_matches)
+      callback(vim.iter(backlink_matches):map(backlink_to_lsp_location):totable())
+    end, opts)
+  end
+
+  if location == "" and (anchor_link or block_link) then
+    local note = api.current_note(0, {
+      collect_anchor_links = anchor_link ~= nil,
+      collect_blocks = block_link ~= nil,
+    })
+    if not note then
+      return log.err "Current buffer does not appear to be a note inside the vault"
+    end
+    find_backlinks(note)
+  elseif anchor_link or block_link then
+    search.resolve_note_async(location, function(notes)
+      find_backlinks(#notes == 1 and notes[1] or nil)
+    end, {
+      notes = {
+        collect_anchor_links = anchor_link ~= nil,
+        collect_blocks = block_link ~= nil,
+      },
+    })
+  else
+    find_backlinks(nil)
+  end
 end
 
-handlers.Wiki = handlers.Markdown
-handlers.WikiWithAlias = handlers.Markdown
-
-handlers.Footnote = function(link, callback)
+local handle_footnote = function(link, callback)
   local footnotes = require "obsidian.footnotes"
   local id = util.parse_link(link)
   assert(id, "failed to parse footnote")
@@ -78,7 +100,7 @@ handlers.Footnote = function(link, callback)
   callback(locations)
 end
 
-handlers.Tag = function(tag, callback)
+local function handle_tag(tag, callback)
   search.find_tags_async(tag, function(tag_locs)
     local lsp_locs = vim.tbl_map(tag_loc_to_lsp_location, tag_locs)
     callback(lsp_locs)
@@ -86,14 +108,14 @@ handlers.Tag = function(tag, callback)
 end
 
 local function collect_current_note(link, link_type, callback)
-  local anchor, block
+  local anchor
+  local block
 
-  if link and link_type == "BlockID" then
-    block = util.parse_link(link)
+  if link and link_type == "block_id" then
+    block = link
   end
 
   -- Check if cursor is on a header, if so and header parsing is enabled, use that anchor.
-  -- TODO: header should just be matched as a type of "ref" by `parse_link`
   if Obsidian.opts.backlinks.parse_headers then
     local header_match = util.parse_header(vim.api.nvim_get_current_line())
     if header_match then
@@ -115,30 +137,27 @@ local function collect_current_note(link, link_type, callback)
   end, { anchor = anchor, block = block })
 end
 
----@type obsidian.search.RefTypes[]
-local supported_reference_types = {
-  Wiki = true,
-  WikiWithAlias = true,
-  Markdown = true,
-  Tag = true,
-  Footnote = true,
-  -- HeaderLink
-  -- BlockLink
-}
-
---- TODO: api.cursor_ref
-
 ---@param include_tag boolean|?
 ---@return string|?
----@return obsidian.search.RefTypes|?
+---@return obsidian.parse.RefKind|"tag"|"block_id"
 local function cursor_ref(include_tag)
   local link, link_type = api.cursor_link()
   if link and link_type then
     return link, link_type
-  elseif include_tag ~= false then
+  end
+
+  local line = vim.api.nvim_get_current_line()
+  local _, cur_col = unpack(vim.api.nvim_win_get_cursor(0))
+  for _, block in ipairs(parse_block_id.extract(line)) do
+    if block.range.start_col <= cur_col and cur_col < block.range.end_col then
+      return block.raw, "block_id"
+    end
+  end
+
+  if include_tag ~= false then
     local tag = api.cursor_tag()
     if tag then
-      return tag, "Tag"
+      return tag, "tag"
     end
   end
 end
@@ -154,16 +173,21 @@ return function(link, opts, callback)
     link, link_type = cursor_ref(opts.tag)
   end
 
-  if not link or not link_type or not supported_reference_types[link_type] then
-    collect_current_note(link, link_type, callback)
-    return
-  end
-
   local wrapped_callback = function(locations)
     callback(nil, locations)
   end
 
-  if handlers[link_type] then
-    handlers[link_type](link, wrapped_callback) -- TODO: maybe same as _definition (location, name, callback)
+  if not link then
+    return collect_current_note(nil, nil, callback)
+  end
+
+  if link_type == "markdown" or link_type == "wiki" then
+    handle_note_ref(link, wrapped_callback)
+  elseif link_type == "footnote" then
+    handle_footnote(link, wrapped_callback)
+  elseif link_type == "tag" then
+    handle_tag(link, wrapped_callback)
+  else -- block id is handled here
+    collect_current_note(link, link_type, callback)
   end
 end