+++ lua/obsidian/lsp/handlers/prepare_rename.lua
@@ -6,9 +6,11 @@ return function(_, handler)
   local link = api.cursor_link()
   local placeholder
   if link then
-    local loc = util.parse_link(link, { strip = true })
+    local loc = util.parse_link(link)
     assert(loc, "wrong link format")
-    placeholder = loc
+    local stripped = util.strip_anchor_links(loc)
+    stripped = util.strip_block_links(stripped)
+    placeholder = stripped ~= "" and stripped or loc
   else
     local note = api.current_note(0)
     assert(note, "not in a obsidian note")