+++ tests/search/test_search.lua
@@ -6,66 +6,6 @@ local new_set, eq = MiniTest.new_set, MiniTest.expect.equality
 
 local T = new_set()
 
-T["find_refs"] = new_set()
-
-T["find_refs"]["should find positions of all refs"] = function()
-  local s = "[[Foo]] [[foo|Bar]]"
-  eq({ { 1, 7, "Wiki", "[[Foo]]" }, { 9, 19, "WikiWithAlias", "[[foo|Bar]]" } }, M.find_refs(s))
-end
-
-T["find_refs"]["should ignore refs within an inline code block"] = function()
-  local s = "`[[Foo]]` [[foo|Bar]]"
-  eq({ { 11, 21, "WikiWithAlias", "[[foo|Bar]]" } }, M.find_refs(s))
-
-  s = "[nvim-cmp](https://github.com/hrsh7th/nvim-cmp) (triggered by typing `[[` for wiki links or "
-    .. "just `[` for markdown links), powered by [`ripgrep`](https://github.com/BurntSushi/ripgrep)"
-  eq({
-    {
-      1,
-      47,
-      "Markdown",
-      "[nvim-cmp](https://github.com/hrsh7th/nvim-cmp)",
-    },
-    {
-      134,
-      183,
-      "Markdown",
-      "[`ripgrep`](https://github.com/BurntSushi/ripgrep)",
-    },
-  }, M.find_refs(s))
-end
-
-T["find_refs"]["should find footnote refs"] = function()
-  local s = "some claim[^1] and [^note]"
-  eq({ { 11, 14, "Footnote", "[^1]" }, { 20, 26, "Footnote", "[^note]" } }, M.find_refs(s))
-end
-
-T["find_refs"]["should find footnote refs in definition lines"] = function()
-  local s = "[^1]: the footnote text"
-  eq({ { 1, 4, "Footnote", "[^1]" } }, M.find_refs(s))
-end
-
-T["find_refs"]["should prefer footnote over markdown link"] = function()
-  local s = "claim[^fn](not a link)"
-  eq({ { 6, 10, "Footnote", "[^fn]" } }, M.find_refs(s))
-end
-
-T["find_refs"]["should not match block wiki links as footnotes"] = function()
-  local s = "[[^block]]"
-  eq({ { 1, 10, "Wiki", "[[^block]]" } }, M.find_refs(s))
-end
-
-T["find_matches"] = function()
-  local matches = M.find_matches(
-    [[
-- <https://youtube.com@Fireship>
-- [Fireship](https://youtube.com@Fireship)
-  ]],
-    { "Markdown" }
-  )
-  eq(1, #matches)
-end
-
 T["find_code_blocks"] = new_set()
 
 T["find_code_blocks"]["should find generic code blocks"] = function()