+++ tests/util/test_parse_link.lua
@@ -9,32 +9,27 @@ T["should parse link"] = function()
   eq(location, "hi#^block")
 end
 
-T["should strip if specified"] = function()
-  local location = M.parse_link("[[hi#^block]]", { strip = true })
-  eq(location, "hi")
-end
-
 T["header link"] = new_set()
 
 T["header link"]["should find in wiki link"] = function()
   local location, name, t = M.parse_link "[[#Header]]"
-  eq(location, "#header")
-  eq(name, "Header")
-  eq(t, "HeaderLink")
+  eq(location, "#Header")
+  eq(name, "#Header")
+  eq(t, "wiki")
 end
 
-T["header link"]["should find in wiki link"] = function()
+T["header link"]["should find in wiki link with alias"] = function()
   local location, name, t = M.parse_link "[[#header|Header]]"
   eq(location, "#header")
   eq(name, "Header")
-  eq(t, "HeaderLink")
+  eq(t, "wiki")
 end
 
 T["header link"]["should find in markdown link"] = function()
   local location, name, t = M.parse_link "[Header](#header)"
   eq(location, "#header")
   eq(name, "Header")
-  eq(t, "HeaderLink")
+  eq(t, "markdown")
 end
 
 T["block link"] = new_set()
@@ -42,22 +37,22 @@ T["block link"] = new_set()
 T["block link"]["should find in wiki link"] = function()
   local location, name, t = M.parse_link "[[#^block]]"
   eq(location, "#^block")
-  eq(name, "block")
-  eq(t, "BlockLink")
+  eq(name, "#^block")
+  eq(t, "wiki")
 end
 
-T["block link"]["should find in wiki link"] = function()
+T["block link"]["should find in wiki link with alias"] = function()
   local location, name, t = M.parse_link "[[#^block|Block]]"
   eq(location, "#^block")
   eq(name, "Block")
-  eq(t, "BlockLink")
+  eq(t, "wiki")
 end
 
 T["block link"]["should find in markdown link"] = function()
   local location, name, t = M.parse_link "[Block](#^block)"
   eq(location, "#^block")
   eq(name, "Block")
-  eq(t, "BlockLink")
+  eq(t, "markdown")
 end
 
 return T